Create Oracle.sol
Demonstrates accessing off-chain data through the oracle pattern. This contract is part of the logic, but most of the logic would have to be either centrally held or managed by a third party such as Oraclize. there are a lot of security risks in this pattern due to import and potential reentrancy.